SkillSubtopicService Completed.

+++ SkillMatrixManagement/src/SkillMatrixManagement.EntityFrameworkCore/Repositories/SkilISubtopicRepository.cs
@@ -6,16 +6,18 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Volo.Abp;
 using Volo.Abp.Domain.Repositories.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore;
+using static SkillMatrixManagement.SkillMatrixManagementDomainErrorCodes;
 
 namespace SkillMatrixManagement.Repositories
 {
-    public class SkilISubtopicRepository : EfCoreRepository<SkillMatrixManagementApplicationDbContext,SkillSubtopic , Guid>, ISkillSubtopicRepository
+    public class SkillSubtopicRepository : EfCoreRepository<SkillMatrixManagementApplicationDbContext,SkillSubtopic , Guid>, ISkillSubtopicRepository
     {
         readonly IDbContextProvider<SkillMatrixManagementApplicationDbContext> _dbContextProvider;
 
-        public SkilISubtopicRepository(IDbContextProvider<SkillMatrixManagementApplicationDbContext> dbContextProvider):base(dbContextProvider) {
+        public SkillSubtopicRepository(IDbContextProvider<SkillMatrixManagementApplicationDbContext> dbContextProvider):base(dbContextProvider) {
         
             _dbContextProvider = dbContextProvider;
         }
@@ -24,7 +26,7 @@ namespace SkillMatrixManagement.Repositories
         {
             if ( skillSubtopic==null)
             {
-                throw new Exception("Invalid Skill content!");
+                throw new BusinessException(SkillMatrixManagementDomainErrorCodes.SkillSubtopicErrorCodes.SkillSubtopicInvalid);
             }
             var dbContext = await _dbContextProvider.GetDbContextAsync();
             var entity = await dbContext.SkillSubtopics.AddAsync(skillSubtopic);
@@ -36,6 +38,13 @@ namespace SkillMatrixManagement.Repositories
         {
             var dbContext = await _dbContextProvider.GetDbContextAsync();
             var entity = await dbContext.SkillSubtopics.FindAsync(skillSubtopicId);
+
+            if (entity == null)
+            {
+                throw new BusinessException(SkillSubtopicErrorCodes.SkillSubtopicNotFoundForDelete);
+                //throw new Exception("id not found!");
+            }
+
             if (entity != null)
             {
                 entity.IsDeleted = true;
@@ -46,20 +55,24 @@ namespace SkillMatrixManagement.Repositories
         public async Task<List<SkillSubtopic>> GetAllAsync()
         {
             var dbContext = await _dbContextProvider.GetDbContextAsync();
-            return await dbContext.SkillSubtopics.ToListAsync();
+            return await dbContext.SkillSubtopics.IgnoreQueryFilters().ToListAsync();
         }
 
         public async Task<SkillSubtopic> GetByIdAsync(Guid id)
         {
             var dbContext = await _dbContextProvider.GetDbContextAsync();
 
-            return await dbContext.SkillSubtopics.FindAsync(id) ?? throw new Exception("ID is not found");
+            return await dbContext.SkillSubtopics.FindAsync(id) ?? throw new BusinessException(SkillMatrixManagementDomainErrorCodes.SkillSubtopicErrorCodes.SkillSubtopicNotFound);
         }
 
         public async Task PermanentDeleteAsync(Guid skillSubtopicId)
         {
             var dbContext = await _dbContextProvider.GetDbContextAsync();
             var entity = await dbContext.SkillSubtopics.FindAsync(skillSubtopicId);
+            if (entity == null)
+            {
+                throw new BusinessException(SkillSubtopicErrorCodes.SkillSubtopicNotFoundForPermanentDelete);
+            }
             if (entity != null)
             {
                 dbContext.SkillSubtopics.Remove(entity);
@@ -71,6 +84,15 @@ namespace SkillMatrixManagement.Repositories
         {
             var dbContext = await _dbContextProvider.GetDbContextAsync();
             var entity = await dbContext.SkillSubtopics.IgnoreQueryFilters().FirstOrDefaultAsync(e => e.Id == skillSubtopicId);
+            if (entity == null)
+            {
+                throw new BusinessException(SkillSubtopicErrorCodes.SkillSubtopicNotFound);
+            }
+            if (!entity.IsDeleted)
+            {
+                throw new BusinessException(SkillSubtopicErrorCodes.SkillSubtopicAlreadyDeleted);
+            }
+
             if (entity != null && entity.IsDeleted)
             {
                 entity.IsDeleted = false;
@@ -90,7 +112,7 @@ namespace SkillMatrixManagement.Repositories
             var entity=await dbContext.SkillSubtopics.FindAsync(skillSubtopic.Id);
             if (entity == null)
             {
-                throw new Exception($"Skill is  not found.");
+                throw new BusinessException(SkillSubtopicErrorCodes.SkillSubtopicNotFoundForUpdate);
             };
             dbContext.Entry(entity).CurrentValues.SetValues(skillSubtopic);
             await dbContext.SaveChangesAsync();
